+++ src/player.h
@@ -32,7 +32,7 @@ public:
 	Deck* get_board() { return &board; };
 	Deck* get_discard() { return &discard; };
 	bool draw();
-	bool cast(unsigned int cardIndex);
+	bool cast(size_t cardIndex);
 };
 
 #endif /* SRC_PLAYER_H_ */
